Rename DockerEventListener extension point and the related stuff
