+++ src/main/java/org/jenkinsci/plugins/docker/traceability/model/DockerTraceabilityReportListener.java
@@ -21,7 +21,7 @@
  * OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN
  * THE SOFTWARE.
  */
-package org.jenkinsci.plugins.docker.traceability;
+package org.jenkinsci.plugins.docker.traceability.model;
 
 import hudson.ExtensionList;
 import hudson.ExtensionPoint;
@@ -29,25 +29,26 @@ import java.util.logging.Level;
 import java.util.logging.Logger;
 import javax.annotation.Nonnull;
 import jenkins.model.Jenkins;
+import org.jenkinsci.plugins.docker.traceability.DockerTraceabilityPlugin;
 import org.jenkinsci.plugins.docker.traceability.fingerprint.DockerDeploymentFacet;
 import org.jenkinsci.plugins.docker.traceability.model.DockerEvent;
 import org.jenkinsci.plugins.docker.traceability.model.DockerTraceabilityReport;
 
 /**
- * An extension point, which allows to subscribe to Docker events in Jenkins.
+ * An extension point, which allows to subscribe to {@link DockerTraceabilityReport}s in Jenkins.
  * @author Oleg Nenashev
  * @since 1.0
  */
-public class DockerEventListener implements ExtensionPoint {
+public class DockerTraceabilityReportListener implements ExtensionPoint {
     
     private final static Logger LOGGER = Logger.getLogger(DockerTraceabilityPlugin.class.getName());
     
     /**
-     * Notifies external listener that an event in Docker happened.
+     * Notifies external listeners that a new Docker report has been received.
      * The method does nothing by default
-     * @param event Event
+     * @param report Event
      */
-    public void onEvent(@Nonnull DockerTraceabilityReport event) {
+    public void onReport(@Nonnull DockerTraceabilityReport report) {
         // Do nothing by default
     }
     
@@ -61,13 +62,13 @@ public class DockerEventListener implements ExtensionPoint {
     }
     
     /**
-     * Fires {@link DockerEvent} on all listeners.
-     * @param event Event to be triggered
+     * Process {@link DockerTraceabilityReport} on all listeners.
+     * @param report Event to be triggered
      */
-    public static void fire(@Nonnull DockerTraceabilityReport event) {
-        for (DockerEventListener listener : all()) {
+    public static void fire(@Nonnull DockerTraceabilityReport report) {
+        for (DockerTraceabilityReportListener listener : all()) {
             try {
-                listener.onEvent(event);
+                listener.onReport(report);
             } catch (Throwable t) { // Prevent failures on runtime exceptions
                 LOGGER.log(Level.SEVERE, "Runtime exception during the event processing in "+ listener, t);
             }
@@ -79,7 +80,7 @@ public class DockerEventListener implements ExtensionPoint {
      * @param containerId Container ID (full 64-char representation)
      */
     public static void fireNewDeployment(@Nonnull String containerId) {
-        for (DockerEventListener listener : all()) {
+        for (DockerTraceabilityReportListener listener : all()) {
             try {
                 listener.onNewDeployment(containerId);
             } catch (Throwable t) { // Prevent failures on runtime exceptions
@@ -90,13 +91,13 @@ public class DockerEventListener implements ExtensionPoint {
 
     /**
      * Retrieves a list of Docker event listeners.
-     * @return A list of all {@link DockerEventListener} extensions.
+     * @return A list of all {@link DockerTraceabilityReportListener} extensions.
      */
-    public static @Nonnull ExtensionList<DockerEventListener> all() {
+    public static @Nonnull ExtensionList<DockerTraceabilityReportListener> all() {
         final Jenkins j = Jenkins.getInstance();
         if (j == null) {
-            return ExtensionList.create((Jenkins) null, DockerEventListener.class);
+            return ExtensionList.create((Jenkins) null, DockerTraceabilityReportListener.class);
         }
-        return j.getExtensionList(DockerEventListener.class);
+        return j.getExtensionList(DockerTraceabilityReportListener.class);
     }
 }
